# connectors/docs/swagger.yaml
definitions:
...
  fiber.Map:
    additionalProperties: {}
    type: object
...
  irminmodels.ConnectorConfigurationValidationResult:
    properties:
      can_connect:
        description: Indicates if the connector can connect to the external system
        example: true
        type: boolean
      connection_details_valid:
        description: Indicates if the connection details are valid
        example: true
        type: boolean
      connection_settings_valid:
        description: Indicates if the connection settings are valid
        example: true
        type: boolean
      errors:
        description: (Optional) Array of validation errors
        example:
        - Invalid host address
        - Authentication failed
        items:
          type: string
        type: array
      ok:
        description: Indicates if the configuration is valid
        example: true
        type: boolean
    type: object
  irminmodels.DynamicField:
    properties:
      default:
        example: Default value
        type: string
      example:
        example: Example value
        maxLength: 100
        type: string
      help_text:
        example: This is a help text
        maxLength: 200
        type: string
      label:
        example: Name
        maxLength: 100
        type: string
      max:
        example: 100
        type: number
      min:
        example: 1
        type: number
      multiple:
        example: false
        type: boolean
      options:
        items:
          $ref: '#/definitions/irminmodels.SelectOption'
        type: array
      required:
        example: true
        type: boolean
      required_with:
        example:
        - field1
        - field2
        items:
          type: string
        type: array
      secret:
        example: false
        type: boolean
      type:
        allOf:
        - $ref: '#/definitions/irminmodels.FieldType'
        enum:
        - text
        - textarea
        - password
        - email
        - checkbox
        - integer
        - float
        - select
        - radio
        - file
        - date
        - time
        - datetime
        example: text
    required:
    - label
    - type
    type: object
  irminmodels.FieldType:
    enum:
    - text
    - textarea
    - password
    - email
    - checkbox
    - integer
    - float
    - select
    - radio
    - file
    - date
    - time
    - datetime
    type: string
    x-enum-varnames:
    - FieldTypeText
    - FieldTypeTextarea
    - FieldTypePassword
    - FieldTypeEmail
    - FieldTypeCheckbox
    - FieldTypeInteger
    - FieldTypeFloat
    - FieldTypeSelect
    - FieldTypeRadio
    - FieldTypeFile
    - FieldTypeDate
    - FieldTypeTime
    - FieldTypeDatetime
...
  irminmodels.SelectOption:
    properties:
      key:
        example: key1
        maxLength: 50
        type: string
      value:
        example: Value 1
        maxLength: 100
        type: string
    required:
    - key
    - value
    type: object
...
host: connectors.irmin.co
info:
  contact:
    email: hello@irmin.co
    name: Irmin Team
    url: https://irmin.co
  description: Collection of connectors for Irmin to interact with external services
  termsOfService: http://irmin.co/terms
  title: Irmin Connectors
  version: 1.0.0
paths:
...
  /sftp/configuration/{key}/fields:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Get dynamic configuration fields for the SFTP connector based on
        the configuration key (details or settings)
      parameters:
      - description: Configuration key
        enum:
        - details
        - settings
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration fields retrieved successfully
          schema:
            additionalProperties:
              $ref: '#/definitions/irminmodels.DynamicField'
            type: object
        "400":
          description: Bad request - invalid configuration key
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Get SFTP connector configuration fields
      tags:
      - sftp
  /sftp/configuration/validate:
    post:
      consumes:
      - multipart/form-data
      description: Validate SFTP connection settings and credentials by testing the
        actual connection
      parameters:
      - description: SFTP server hostname or IP address
        in: formData
        name: details[host]
        required: true
        type: string
      - description: 'SFTP server port (default: 22)'
        in: formData
        name: details[port]
        type: integer
      - description: Username for SFTP authentication
        in: formData
        name: details[username]
        required: true
        type: string
      - description: Password for SFTP authentication (if not using private key)
        in: formData
        name: details[password]
        type: string
      - description: Private key for SFTP authentication (if not using password)
        in: formData
        name: details[private_key]
        type: string
      - description: Passphrase for encrypted private key
        in: formData
        name: details[private_key_passphrase]
        type: string
      - description: Expected host key fingerprint for security verification
        in: formData
        name: details[host_key_fingerprint]
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration validation result
          schema:
            $ref: '#/definitions/irminmodels.ConnectorConfigurationValidationResult'
        "400":
          description: Bad request - invalid configuration data
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Validate SFTP connector configuration
      tags:
      - sftp
...
  /stripe/configuration/{key}/fields:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Get dynamic configuration fields for the Stripe connector based
        on the configuration key (details or settings)
      parameters:
      - description: Configuration key
        enum:
        - details
        - settings
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Configuration fields retrieved successfully
          schema:
            additionalProperties:
              $ref: '#/definitions/irminmodels.DynamicField'
            type: object
        "400":
          description: Bad request - invalid configuration key
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized - invalid or missing authentication
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - SystemTokenAuth: []
      summary: Get Stripe connector configuration fields
      tags:
      - stripe
...
schemes:
- https
securityDefinitions:
  OperationTokenAuth:
    description: Operation token with Bearer prefix - received from operation/init
      and used for data operation endpoints
    in: header
    name: Authorization
    type: apiKey
  SystemTokenAuth:
    description: System token with Bearer prefix - used for connector registration
      and initialization endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"